+++ R/STMF.R
@@ -33,32 +33,32 @@ read_stmf <- function(country) {
     country,
     "stmfout.csv"
   )
-  read_stmf_file(url)
+  read_stmf_files(url)
 }
 
 #' Read STMF data from files downloaded from HMD
 #'
-#' `read_stmf_file` reads weekly mortality data from a file downloaded
+#' `read_stmf_files` reads weekly mortality data from a file downloaded
 #' from the Short-term Mortality Fluctuations (STMF) series available in the
 #' Human Mortality Database (HMD) <https://www.mortality.org/Data/STMF>),
 #' and constructs a `vital` object suitable for use in other functions.
 #'
 #' @param file Name of a file containing data downloaded from the HMD.
 #'
-#' @return `read_stmf_file` returns a `vital` object combining the downloaded data.
+#' @return `read_stmf_files` returns a `vital` object combining the downloaded data.
 #'
 #' @author Rob J Hyndman
 #' @examples
 #' \dontrun{
 #' # File downloaded from the [Human Mortality Database STMF series]
 #' (https://www.mortality.org/Data/STMF)
-#' mortality <- read_stmf_file("AUSstmfout.csv")
+#' mortality <- read_stmf_files("AUSstmfout.csv")
 #' }
 #' @keywords manip
 #' @export
 #'
 
-read_stmf_file <- function(file) {
+read_stmf_files <- function(file) {
   data <- utils::read.csv(
     file,
     header = TRUE,