+++ NEWS.md
@@ -1,6 +1,6 @@
 # vital (development version)
 
-* New data functions: read_ktdb(), read_ktdb_file(), read_stmf(), read_stmf_file()
+* New data functions: read_ktdb(), read_ktdb_files(), read_stmf(), read_stmf_files()
 * New model functions: GAPC(), PLAT(), M7(), RH(), CBD(), LC2()
 * New smoothing function: smooth_mortality_law()
 * New stochastic simulation function: generate_population()
@@ -18,5 +18,3 @@
 # vital 1.0.0
 
 * First CRAN submission
-
-